operation: change nonce from u8 to u64
rationale: in eltoo the graph can grow up to u48 transactions. Even though not of all would be published, some may, and all of them must be ordered appropriately

+++ src/vm/contract.rs
@@ -215,7 +215,7 @@ impl<'op> Operation for OrdOpRef<'op> {
         }
     }
 
-    fn nonce(&self) -> u8 {
+    fn nonce(&self) -> u64 {
         match self {
             OrdOpRef::Genesis(op) => op.nonce(),
             OrdOpRef::Transition(op, ..) => op.nonce(),
@@ -419,14 +419,14 @@ pub enum OpOrd {
     Extension {
         witness: WitnessOrd,
         // TODO: Consider using extension type here
-        nonce: u8,
+        nonce: u64,
         opid: OpId,
     },
     #[strict_type(tag = 0xFF)]
     Transition {
         witness: WitnessOrd,
         // TODO: Consider using transition type here
-        nonce: u8,
+        nonce: u64,
         opid: OpId,
     },
 }
@@ -468,7 +468,7 @@ impl GlobalOrd {
             idx,
         }
     }
-    pub fn transition(opid: OpId, idx: u16, nonce: u8, witness: WitnessOrd) -> Self {
+    pub fn transition(opid: OpId, idx: u16, nonce: u64, witness: WitnessOrd) -> Self {
         Self {
             op_ord: OpOrd::Transition {
                 witness,
@@ -478,7 +478,7 @@ impl GlobalOrd {
             idx,
         }
     }
-    pub fn extension(opid: OpId, idx: u16, nonce: u8, witness: WitnessOrd) -> Self {
+    pub fn extension(opid: OpId, idx: u16, nonce: u64, witness: WitnessOrd) -> Self {
         Self {
             op_ord: OpOrd::Extension {
                 witness,
